파일 정리 및 Readme 추가
