+++ 1. Preparing Corpus/jeongmin/call_rate/call_rate/items.py
@@ -1,0 +1,13 @@
+# Define here the models for your scraped items
+#
+# See documentation in:
+# https://docs.scrapy.org/en/latest/topics/items.html
+
+import scrapy
+
+
+class CallRateItem(scrapy.Item):
+    # define the fields for your item here like:
+    date = scrapy.Field()
+    callrate = scrapy.Field()
+
