+++ 1. Preparing Corpus/jeongmin/call_rate/scrapy.cfg
@@ -1,0 +1,11 @@
+# Automatically created by: scrapy startproject
+#
+# For more information about the [deploy] section see:
+# https://scrapyd.readthedocs.io/en/latest/deploy.html
+
+[settings]
+default = call_rate.settings
+
+[deploy]
+#url = http://localhost:6800/
+project = call_rate
